backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pyOpenSSL==23.3.0
python-dotenv==1.0.1
werkzeug==3.0.1
Flask-Mail==0.10.0
